+++ lib/config/bloc_provider/register_reusable_editor_bloc_provider.dart
@@ -7,20 +7,28 @@ import 'package:reusable_editor/reusable_editor.dart';
 
 late TextFieldCubit _textFieldCubit;
 late DateTimeCubit _dateTimeCubit;
-late ImageFieldCubit _imageCrudCubit;
 late SwitchCubit _switchCubit;
 late EnumOptionCubit _selectionCubit;
 late ToggleCubit _toggleCubit;
+late VideoFieldCubit _videoFieldCubit;
+late AudioFieldCubit _audioFieldCubit;
+late DocFieldCubit _docFieldCubit;
+late ImageFieldCubit _imageFieldCubit;
+late ListImageFieldCubit _listImageFieldCubit;
 // late SearchFieldCubit _searchFieldCubit;
 
 initReusableEditorBlocProvider() {
   // ---- Reusable Editor Init ------
   _textFieldCubit = sl<TextFieldCubit>();
   _dateTimeCubit = sl<DateTimeCubit>();
-  _imageCrudCubit = sl<ImageFieldCubit>();
   _switchCubit = sl<SwitchCubit>();
   _selectionCubit = sl<EnumOptionCubit>();
   _toggleCubit = sl<ToggleCubit>();
+  _videoFieldCubit = sl<VideoFieldCubit>();
+  _audioFieldCubit = sl<AudioFieldCubit>();
+  _docFieldCubit = sl<DocFieldCubit>();
+  _imageFieldCubit = sl<ImageFieldCubit>();
+  _listImageFieldCubit = sl<ListImageFieldCubit>();
   // _searchFieldCubit = sl<SearchFieldCubit>();
 }
 
@@ -28,10 +36,14 @@ disposeReusableEditorBlocProvider() {
   // ---- Reusable Editor Dispose ------
   _textFieldCubit.close();
   _dateTimeCubit.close();
-  _imageCrudCubit.close();
   _switchCubit.close();
   _selectionCubit.close();
   _toggleCubit.close();
+  _videoFieldCubit.close();
+  _audioFieldCubit.close();
+  _docFieldCubit.close();
+  _imageFieldCubit.close();
+  _listImageFieldCubit.close();
   // _searchFieldCubit.close();
 }
 
@@ -39,9 +51,13 @@ List<SingleChildWidget> reusableEditorBlocProviders = [
   // ---- Reusable Editor Bloc Provider ------
   BlocProvider<TextFieldCubit>(create: (_) => _textFieldCubit),
   BlocProvider<DateTimeCubit>(create: (_) => _dateTimeCubit),
-  BlocProvider<ImageFieldCubit>(create: (_) => _imageCrudCubit),
   BlocProvider<SwitchCubit>(create: (_) => _switchCubit),
   BlocProvider<EnumOptionCubit>(create: (_) => _selectionCubit),
   BlocProvider<ToggleCubit>(create: (_) => _toggleCubit),
+  BlocProvider<VideoFieldCubit>(create: (_) => _videoFieldCubit),
+  BlocProvider<AudioFieldCubit>(create: (_) => _audioFieldCubit),
+  BlocProvider<DocFieldCubit>(create: (_) => _docFieldCubit),
+  BlocProvider<ImageFieldCubit>(create: (_) => _imageFieldCubit),
+  BlocProvider<ListImageFieldCubit>(create: (_) => _listImageFieldCubit),
   // BlocProvider<SearchFieldCubit>(create: (_) => _searchFieldCubit),
 ];